I've organized the codebase into separate files based on class responsibility:
- Configuration classes (BotConfig, TokenConfig) are now in `config.py`.
- Data models (Token, Wallet, Transaction) have been moved to `models.py`.
- DatabaseManager is now located in `database.py`.
- LedgerManager has been moved to `ledger.py`.
- UserStates can now be found in `bot_states.py`.
- I've updated `main.py` to import from these new modules and removed the original class definitions.
- I also made sure all import paths are correct and dependencies are properly managed.
