+++ config.py
@@ -1,0 +1,12 @@
+class BotConfig:
+    BOT_TOKEN = "7861260810:AAEuI-huruUCJyNPgnMbck2t2AnY4pJejD8"
+    ADMIN_IDS = [6328016694]
+    DB_PATH = "helgykoin.db"
+
+class TokenConfig:
+    NAME = "HelgyKoin"
+    SYMBOL = "HKN"
+    DECIMALS = 8
+    TOTAL_SUPPLY = 1_000_000_000.0
+    INITIAL_PRICE = 0.0001
+    STARTUP_BONUS = 100.0
